Split cameraman into example and logic modules
Remove the monolithic Elements/extensions/camera_man/cameraman.py and split its functionality into two files: cameraman_example.py (example usage / scene setup) and cameraman_logic.py (camera UI, path logic and runtime functions). The refactor modularizes UI and camera update logic, exposes run_camera and setter functions (set_control_points, set_target, set_up, set_cam_pos) for external control, and adjusts orbit/bezier calculations to accept a target offset. This makes the camera example reusable and easier to maintain.
